backend/requirements.txt
# R.O.N.D.A. Backend — Python 3.10
Django==5.2.11
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
psycopg[binary]==3.3.3
Pillow==12.1.1
